kentik_api_library/requirements.txt
astroid>=2.5.6
attrs>=21.2.0
certifi>=2020.12.5
chardet>=4.0.0
dacite>=1.6.0
fastparquet>=0.6.3
GitPython>=3.1.24
grpcio>=1.38.1
httpretty>=1.1.3
idna>=2.10
iniconfig>=1.1.1
lazy-object-proxy>=1.6.0
mccabe>=0.6.1
packaging>=20.9
pluggy>=0.13.1
protobuf>=3.19.1
py>=1.10.0
pyparsing>=2.4.7
python-http-client>=3.3.2
requests[socks]>=2.25.1
six>=1.16.0
toml>=0.10.2
traitlets>=5.0.5
typed-ast>=1.4.3
typing-extensions>=3.10.0.0
urllib3>=1.26.4
wheel>=0.36.2
wrapt>=1.12.1
